Inject HttpClient so that its configuration can be overridden.
It's not enough to inject the SSLSocketFactory since that is not
the only thing that needs to be configured in a HttpClient. When
running in Eclipse for instance, the HttpClient must be configured
with HttpRoutePlanner that takes Eclipse configured proxies into
account.

This commit ensures that both the SSLSocketFactory and the HttpClient
are accessed by injection. It also adds a default binding for the
HttpClient that preserves the old behavior.